[TASK] DatabaseBackend writes account info
Since the user action logger was replaced with the account action
logger the database backend now stores the account identifier and the
authentication provider name. This makes it possible to see which
account trigger the action, even after the account was deleted.

The log entry domain model was modified to reflect the changes. It has
two new properties for storing the account information. The user setter
was replaced with an account setter and the user is now optional
(nullable).